requirements.txt
openai>=1.0.0
anthropic>=0.20.0
python-dotenv>=1.0.0
pandas>=2.0.0
colorama>=0.4.6
tabulate>=0.9.0

# ── Task 2: RAG Document Analysis ────────────────────────────────────────────
langchain>=0.2.0
langchain-openai>=0.1.8
langchain-community>=0.2.0
faiss-cpu>=1.7.4
matplotlib>=3.7.0
numpy>=1.24.0
kagglehub>=0.2.0